Moves and fixes tests
